Basic perf update
