+++ expressions2/operator_node.cc
@@ -1,11 +1,5 @@
 #include "operator_node.h"
 
-#include <glog/logging.h>
-
-#include <cmath>
-
-#include "consts.h"
-
 namespace code_experiments {
 
 template <typename T>
@@ -16,18 +10,6 @@ OperatorNode<T>::OperatorNode(
     : left_(left),
       operator_(op),
       right_(right) {}
-
-template <typename T>
-OperatorNode<T>::OperatorNode(std::istream &in, const std::string &specifier)
-    : operator_('?') {
-  if (specifier == kTOperatorStr) {
-    ternary_ = std::move(Expression<T>(in));
-  } else {
-    in >> operator_;
-  }
-  left_ = std::move(Expression<T>(in));
-  right_ = std::move(Expression<T>(in));
-}
 
 template <typename T>
 OperatorNode<T>::OperatorNode(
@@ -54,28 +36,6 @@ OperatorNode<T>::OperatorNode(const OperatorNode<T> &from)
     : left_(from.left_),
       operator_(from.operator_),
       right_(from.right_) {}
-
-template <typename T>
-T OperatorNode<T>::Eval() {
-  T left = left_.Eval();
-  T right = right_.Eval();
-  switch (operator_) {
-    case '+':
-      return left + right;
-    case '-':
-      return left - right;
-    case '*':
-      return left * right;
-    case '/':
-      return left / right;
-    case '^':
-      return std::pow(left, right);
-    case '?':
-      return ternary_.Eval() != 0 ? left : right;
-    default:
-      LOG_ASSERT(false) << "Unexpected operator: " << operator_;
-  }
-}
 
 template <typename T>
 void OperatorNode<T>::PrintAsTree(int indent) {
